+++ src/Exception/IntegerOverflowException.php
@@ -23,6 +23,6 @@ final class IntegerOverflowException extends MathException
     {
         $message = '%s is out of range %d to %d and cannot be represented as an integer.';
 
-        return new self(sprintf($message, (string) $value, PHP_INT_MIN, PHP_INT_MAX));
+        return new self(sprintf($message, $value->toString(), PHP_INT_MIN, PHP_INT_MAX));
     }
 }