+++ tests/BigIntegerTest.php
@@ -4348,7 +4348,7 @@ class BigIntegerTest extends AbstractTestCase
     #[DataProvider('providerModInverse')]
     public function testModInverse(string $x, BigNumber|int|float|string $m, string $expectedResult): void
     {
-        self::assertSame($expectedResult, (string) BigInteger::of($x)->modInverse($m));
+        self::assertSame($expectedResult, BigInteger::of($x)->modInverse($m)->toString());
     }
 
     public static function providerModInverse(): array
@@ -4972,7 +4972,7 @@ class BigIntegerTest extends AbstractTestCase
     public function testFromBytes(string $byteStringHex, bool $signed, string $expectedNumber): void
     {
         $number = BigInteger::fromBytes(hex2bin($byteStringHex), $signed);
-        self::assertSame($expectedNumber, (string) $number);
+        self::assertSame($expectedNumber, $number->toString());
     }
 
     public static function providerFromBytes(): Generator
@@ -5351,6 +5351,27 @@ class BigIntegerTest extends AbstractTestCase
         $value = '123456789123456789123456789123456789';
         $random = BigInteger::randomRange($value, $value);
         self::assertBigIntegerEquals($value, $random);
+    }
+
+    #[DataProvider('providerToString')]
+    public function testToString(int|float|string $value, string $expected): void
+    {
+        $bigInteger = BigInteger::of($value);
+        self::assertSame($expected, $bigInteger->toString());
+        self::assertSame($expected, (string) $bigInteger);
+    }
+
+    public static function providerToString(): array
+    {
+        return [
+            [-1, '-1'],
+            [0, '0'],
+            [1, '1'],
+            [1e2, '100'],
+            [-1e3, '-1000'],
+            ['7349837498278937849739739878293798239223', '7349837498278937849739739878293798239223'],
+            ['-7349837498278937849739739878293798239223', '-7349837498278937849739739878293798239223'],
+        ];
     }
 
     public function testSerialize(): void