+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 ✨ **New features**
 
 - New method: `BigInteger::isBitSet()` (replaces `testBit()`)
+- New method: `BigNumber::toString()` (alias of magic method `__toString()`)
 
 ## [0.14.7](https://github.com/brick/math/releases/tag/0.14.7) - 2026-02-07